+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Select status on `issue create` and `template evaluate`
+
 ## 2023-09-23 v0.1.7
 
 - Add arguments to `issue list`